Split DGUSDisplay and DGUSScreenVariableHandler
