+++ Marlin/src/lcd/extui/lib/dgus/hiprecy/DGUSDisplayDef.cpp
@@ -28,6 +28,7 @@
 
 #include "../DGUSDisplayDef.h"
 #include "../DGUSDisplay.h"
+#include "../DGUSScreenVariableHandler.h"
 
 #include "../../../../../module/temperature.h"
 #include "../../../../../module/motion.h"
